// superset-frontend/src/features/posts/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

export type Role = {
  id: number;
  name: string;
};

export type PostObject = {
  id: number;
  name: string;
  label?: string;
  description?: string;
  active: boolean;
  role_ids: number[];
  role_names: string[];
  user_ids: number[];
  users: { id: number; username: string }[];
  user_count: number;
};

export type UserOption = {
  value: number;
  label: string;
};

export interface FormValues {
  name: string;
  label?: string;
  description?: string;
  active?: boolean;
  roles: number[];
  users: UserOption[];
}

export interface PostModalProps {
  show: boolean;
  onHide: () => void;
  onSave: () => void;
  roles: Role[];
  isEditMode?: boolean;
  post?: PostObject;
}